:art: Store restructured and header is now fixed to the top of the screen
