+++ src/routes.js
@@ -1,6 +1,6 @@
 import Home from './views/Home.vue';
 import About from './views/AboutUs.vue';
-import Dummy from './views/Dummy.vue'
+import Dummy from './views/Dummy.vue';
 
 export default [
    { name: 'home', path: '/', component: Home },
