+++ src/store/index.js
@@ -1,4 +1,4 @@
-import Vuex from 'vuex'
+ import Vuex from 'vuex'
 import {store as articles} from './modules/articles/store'
 import {store as navigation} from './modules/navigation/store'
 import {store as pages} from './modules/pages/store'
